Delete instance directory
The instance/study.db database has so far served as a for-development dummy database for testing purposes. Now that a production database with real information will be used, this file will be removed from the repo (frankly, I should have used .gitignore to keep it out of the repo in the first place).